I added this last time but forgetted it this time
You need ctime for your srand to work. Also, maybe I'm wrong, but sstream doesn't appear in our main. Also, for convention, I'd not put iostream or namespace std in a header file. Just because you'd define everything that needs iostream in the class file, and using "std::" is convention for headers. Otherwise, we's good.